PRM init working with obstacle detection + astar tested
