+++ ProjetSupaero2018/catkin_ws/src/pie86_obstacles/scripts/check_point_client.py
@@ -2,7 +2,7 @@
 
 import sys
 import rospy
-from obstacles.srv import *
+from pie86_obstacles.srv import *
 
 global SERVICE_NAME
 SERVICE_NAME = 'check_point'
