+++ ProjetSupaero2018/catkin_ws/src/pie86_obstacles/scripts/read_obstacles_server.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python
 
-from obstacles.srv import *
+from pie86_obstacles.srv import *
 import rospy
 import obstacles_functions
 
 def handle_obstacles(req):
     print "Returning obstacles from file [%s]"%(req.file)
     vec, size = obstacles_functions.read_obstacles_function(req.file)
-    
+
     print "Obstacles of size %s are : [%s]"%(size,vec)
     return ReadObstaclesResponse(size, vec)
