+++ ProjetSupaero2018/catkin_ws/src/roadmap/launch/init_prm.launch
@@ -1,0 +1,9 @@
+<!-- launches Hector Quadrotor simulation and IntelTeleop optimal control nodes -->
+<launch>
+
+  <!-- <arg name="init" default="true"/> -->
+  <node pkg="pie86_obstacles" type="check_point_server.py" name="NameCheckPoint"/>
+
+  <node pkg="roadmap" type="init_prm.py" name="NameRoadMap"/>
+
+</launch>
