+++ ProjetSupaero2018/catkin_ws/src/roadmap/scripts/init_prm.py
@@ -1,24 +1,39 @@
+#!/usr/bin/env python
+import os
 import random
-import numpy as np
-import heapq
 
-import rospy
-from obstacles.srv import *
+try:
+    import rospy
+    import rospkg
+    from pie86_obstacles.srv import *
+except:
+    print('ROS not runnning')
+
+import prm
 
 CHECKPOINT_SERVICE = 'check_point'
+PACKAGE_NAME = 'roadmap'
 
+rospackage = rospkg.RosPack()
+PACKAGE_DATA_PATH = os.path.join(rospackage.get_path(PACKAGE_NAME), 'data')
+if not os.path.exists(PACKAGE_DATA_PATH):
+    os.makedirs(PACKAGE_DATA_PATH)
+
+NB_SAMPLE = 100
+NB_CONECT = 2
 # node: index of a node
 # state: tuple representing the state associated or not with a node
 
 def main():
     rospy.init_node('init_the_prm')
-    rospy.wait_for_service(CHECKPOINT_SERVICE)
+    rospy.loginfo('Init PRM initialized')
+    rospy.wait_for_service('check_point')
+    rospy.loginfo('End of wait for check_point')
 
-    state_space = StateSpace([(0, 10), (0, 5)])
-    nb_sample = 10
-    nb_connect = 3
-    graph = prm_init(state_space, nb_sample, nb_connect)
-    print(graph)
+    state_space = prm.StateSpace([(0, 10), (0, 5)], seed=0)
+    graph = prm_init(state_space, NB_SAMPLE, NB_CONECT)
+    # rospy.loginfo(graph)
+    graph.save(os.path.join(PACKAGE_DATA_PATH))
 
 
 def prm_init(state_space, nb_sample, nb_connect, nb_best=None):
@@ -28,7 +43,7 @@ def prm_init(state_space, nb_sample, nb_connect, nb_best=None):
     nb_best: number of the best edges to keep
     """
 
-    graph = Graph(state_space, euclid)
+    graph = prm.Graph(state_space, prm.euclid)
     for _ in range(nb_sample):
         node, s = graph.new_node()
         if valid_state(s):
@@ -40,10 +55,12 @@ def prm_init(state_space, nb_sample, nb_connect, nb_best=None):
         new_edges = []
         for close_node in closest_nodes:
             # !! with Acado, to be done in the 2 ways !!
-            conn = connection(node, close_node)
+            s = graph.nodes[node][0]
+            close_s = graph.nodes[close_node][0]
+            conn = connection(s, close_s)
             if conn[0]:
                 new_edges.append((node, close_node, conn[1]))
-            conn = connection(close_node, node)
+            conn = connection(close_s, s)
             if conn[0]:
                 new_edges.append((close_node, node, conn[1]))
 
@@ -58,7 +75,7 @@ def prm_init(state_space, nb_sample, nb_connect, nb_best=None):
 
 def valid_state(s):
     try:
-        get_if_valid = rospy.ServiceProxy(SERVICE_NAME, CheckPoint)
+        get_if_valid = rospy.ServiceProxy(CHECKPOINT_SERVICE, CheckPoint)
         resp = get_if_valid(*s)
         return resp.is_valid
     except rospy.ServiceException, e:
@@ -66,132 +83,10 @@ def valid_state(s):
     # return s[0] + s[1] < 8 or s[0] + s[1] > 10
 
 
-def connection(node1, node2):
+def connection(s1, s2):
     # dumb function, return a false
-    return (True, random.random())
+    return (True, prm.euclid(s1, s2))
 
-
-class StateSpace:
-    def __init__(self, bounds, seed=None):
-        self.bounds = bounds
-        random.seed(seed)
-
-    def rdm_state(self):
-
-        return [b[0] + random.random()*(b[1] - b[0]) for b in self.bounds]
-
-
-class Graph:
-    def __init__(self, state_space, hdistance):
-        self.state_space = state_space
-        self.hdistance = hdistance
-
-        # idx_node: (state, linked_nodes)
-        # ex: 4: ((0.56, 4.2), [4, 12, 5])
-        self.nodes = {}
-        # (idx_node1, idx_node2): distance
-        # ex: (4, 2): 2.5
-        self.edges = {}
-
-    def __str__(self):
-        return 'Nodes: ' + str(self.nodes) + '\n\n' + 'Edges: ' + str(self.edges)
-
-    def save(self, path):
-        pass
-
-    def load(self, path):
-        pass
-
-    def new_node(self):
-        return len(self.nodes), self.state_space.rdm_state()
-
-    def add_node(self, idx, state):
-        self.nodes[idx] = (state, [])
-
-    def add_edge(self, edge):
-        self.nodes[edge[0]][1].append(edge[1])
-        self.nodes[edge[1]][1].append(edge[0])
-        self.edges[(edge[0], edge[1])] = edge[2]
-
-    def closest_nodes(self, state, nb_connect, new=True):
-        """
-        new bool: enables to avoid comparing the new node with itself if True
-        """
-        # -1 to avoid comparing the new node with itself but dirty
-        existing_nodes = list(range(len(self.nodes) - 1*new))
-        distances = [self.hdistance(state, self.nodes[node_g][0])
-                     for node_g in existing_nodes]
-        nb_connect = min(nb_connect, len(distances))
-        # argpartition: return a list of node indices, the n first of which
-        # correspond to the n lowest distance unsorted
-        if len(distances) <= nb_connect:
-            return existing_nodes
-
-        return np.argpartition(distances, nb_connect)[:nb_connect]
-
-
-def astar(start, goal, graph, hdistance):
-    """Find the shortest between two nodes in the graph"""
-    if start not in graph.nodes:
-        raise ValueError('node ' + str(start) + ' not in the graph')
-    if goal not in graph.nodes:
-        raise ValueError('node ' + str(goal) + ' not in the graph')
-
-    # list of (node index, heuristic)
-    frontier = PriorityQueue()
-    frontier.put(start, 0)
-    came_from = {}
-    cost_so_far = {}
-    came_from[start] = None
-    cost_so_far[start] = 0
-
-    while not frontier.empty():
-        current = frontier.get()
-        print('came_from')
-        print(came_from)
-        print('cost_so_far')
-        print(cost_so_far)
-        if current == goal:
-            # path reconstitution starting from the end
-            shortest_path = [goal]
-            while shortest_path[-1] != start:
-                shortest_path.append(came_from[shortest_path[-1]])
-
-            return list(reversed(shortest_path))
-        for neigh in graph.nodes[current][1]:
-
-            new_cost = cost_so_far[current] + graph.edges[(current, neigh)]
-
-            if neigh not in cost_so_far or new_cost < cost_so_far[neigh]:
-                cost_so_far[neigh] = new_cost
-                dist_goal = hdistance(graph.nodes[current][0], graph.nodes[goal][0])
-                priority = new_cost + dist_goal
-                frontier.put(neigh, priority)
-                came_from[neigh] = current
-
-    # no path found
-    return None
-
-
-class PriorityQueue:
-    def __init__(self):
-        self.elements = []
-
-    def empty(self):
-        return len(self.elements) == 0
-
-    def put(self, item, priority):
-        heapq.heappush(self.elements, (priority, item))
-
-    def get(self):
-        return heapq.heappop(self.elements)[1]
-
-def euclid(s1, s2):
-    return np.sqrt((s2[0]-s1[0])**2 + (s2[1]-s1[1])**2)
 
 if __name__ == '__main__':
-    state_space = StateSpace([(0, 10), (0, 5)])
-    nb_sample = 10
-    nb_connect = 3
-    graph = prm_init(state_space, nb_sample, nb_connect)
-    print(graph)
+    main()
